Consegna esercitazione 4
aggiornamento parte di codice

+++ esercitazione4/is_sorted.h
@@ -9,7 +9,7 @@ bool is_sorted(const std::vector<T>& vec)
     if (vec.size() == 0) {
         return true;
     } else {
-        for (size_t i = 1; i < vec.size(); i++) {
+        for (int i = 1; i < vec.size(); i++) {
         if (vec[i] < vec[i - 1]) {
             return false;
         }
